+++ OpenResKit.Energy/Reading.cs
@@ -21,7 +21,7 @@ namespace OpenResKit.Energy
   public class Reading
   {
     public virtual int Id { get; set; }
-    public virtual long CounterReading { get; set; }
+    public virtual double CounterReading { get; set; }
     public virtual DateTime ReadingDate { get; set; }
     public string MeasuringDevice { get; set; }
   }
